@@ -29,39 +30,45 @@
     "clean": "rimraf ./dist && rimraf ./node_modules && npm cache clean --force && git checkout -- package.json && git checkout -- package-lock.json",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0",
     "changelog-current": "conventional-changelog -p angular -r 2",
+    "preversion.linux": "npm run test.all && rimraf ./dist && npm run build.prod",
+    "version.linux": "npm run changelog && git add -A CHANGELOG.md",
+    "postversion.linux": "git push && git push --tags && rimraf igo2.zip && cd ./dist/igo2/ && zip -r ../igo2.zip ./* && npm run ghpages && ghrelease",
     "preversion": "npm run test.all && rimraf ./dist && npm run build.prod",
     "version": "npm run changelog && git add -A CHANGELOG.md",
-    "postversion": "git push && git push --tags && rimraf igo2.zip && cd ./dist/igo2/ && zip -r ../igo2.zip ./* && npm run ghpages && npm run ghrelease",
-    "preghpages": "npm run build -- --configuration=github --output-path ./dist/ghpages --base-href /igo2/ && echo include: ['_default.json', '_contexts.json', '_base.json'] > dist/ghpages/_config.yml",
-    "ghpages": "npx ngh --dir=dist/ghpages --no-silent=false --message=$(node -p -e \"require('./package.json').version\")",
+    "postversion": "git push && git push --tags && rimraf igo2.zip && cd ./dist/igo2/ && tar -a -c -f ../igo2.zip * && npm run ghpages.win && ghrelease.win",
+    "preghpages": "npm run build -- --configuration=github --output-path ./dist/ghpages --base-href /igo2/ && echo \"include: ['_default.json', '_contexts.json', '_base.json']\" > dist/ghpages/_config.yml",
+    "ghpages": "VERSION=$(node -p -e \"require('./package.json').version\") && npx ngh --dir=dist/ghpages --no-silent=false --message=$VERSION",
     "postghpages": "rimraf ./dist/ghpages",
-    "ghrelease": "publish-release --owner infra-geo-ouverte --repo igo2 --tag $(node -p -e \"require('./package.json').version\") --name $(node -p -e \"require('./package.json').version\") --reuse-release --notes \"$(npm run changelog-current -s)\" --assets dist/igo2.zip"
+    "preghpages.win": "npm run build -- --configuration=github --output-path ./dist/ghpages --base-href /igo2/ && echo include: ['_default.json', '_contexts.json', '_base.json'] > dist/ghpages/_config.yml",
+    "ghpages.win": "npx ngh --dir=dist/ghpages --no-silent=false --message=%npm_package_version%",
+    "ghrelease": "VERSION=$(node -p -e \"require('./package.json').version\") && publish-release --owner infra-geo-ouverte --repo igo2 --tag $VERSION --name $VERSION --reuse-release --notes \"$(npm run changelog-current -s)\" --assets dist/igo2.zip",
+    "preghrelease.win": "rimraf igo2.zip && cd ./dist/igo2/ && tar -a -c -f ../igo2.zip *",
+    "ghrelease.win": "gh-release -o infra-geo-ouverte -r igo -t %npm_package_version% -n %npm_package_version% -a dist/igo2.zip"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.1.3",
-    "@angular/cdk": "^12.1.3",
-    "@angular/common": "^12.1.3",
-    "@angular/compiler": "^12.1.3",
-    "@angular/core": "^12.1.3",
-    "@angular/forms": "^12.1.3",
-    "@angular/material": "^12.1.3",
-    "@angular/platform-browser": "^12.1.3",
-    "@angular/platform-browser-dynamic": "^12.1.3",
-    "@angular/router": "^12.1.3",
+    "@angular/animations": "^12.2.9",
+    "@angular/cdk": "^12.2.9",
+    "@angular/common": "^12.2.9",
+    "@angular/compiler": "^12.2.9",
+    "@angular/core": "^12.2.9",
+    "@angular/forms": "^12.2.9",
+    "@angular/material": "^12.2.9",
+    "@angular/platform-browser": "^12.2.9",
+    "@angular/platform-browser-dynamic": "^12.2.9",
+    "@angular/router": "^12.2.9",
     "@azure/msal-angular": "^2.0.1",
     "@azure/msal-browser": "^2.15.0",
-    "@igo2/angular2-notifications": "^12.0.1",
-    "@igo2/auth": "~1.9.0",
-    "@igo2/common": "~1.9.0",
-    "@igo2/context": "~1.9.0",
-    "@igo2/core": "~1.9.0",
-    "@igo2/geo": "~1.9.0",
-    "@igo2/integration": "~1.9.0",
-    "@igo2/utils": "~1.9.0",
+    "@igo2/auth": "~1.10.0",
+    "@igo2/common": "~1.10.0",
+    "@igo2/context": "~1.10.0",
+    "@igo2/core": "~1.10.0",
+    "@igo2/geo": "~1.10.0",
+    "@igo2/integration": "~1.10.0",
+    "@igo2/utils": "~1.10.0",
     "@mat-datetimepicker/core": "^7.0.1",
-    "@turf/buffer": "^5.1.5",
-    "@turf/point-on-feature": "^5.1.5",
+    "@turf/buffer": "^6.5.0",
+    "@turf/point-on-feature": "^6.5.0",
     "angular-shepherd": "12.0.0",
     "bowser": "^2.10.0",
     "classlist.js": "^1.1.20150312",
@@ -121,12 +137,10 @@
     "ng-packagr": "^12.1.2",
     "npx": "^10.2.0",
     "protractor": "~7.0.0",
-    "publish-release": "^1.6.1",
-    "sass": "^1.37.5",
+    "sass": "^1.42.1",
     "ts-node": "^10.1.0",
-    "tslint": "~6.1.0",
     "typescript": "~4.3.5",
     "webdriver-manager": "^12.1.6",
-    "webpack": "^4.41.2"
+    "webpack": "^5.58.1"
   }
 }
