@@ -8,8 +8,8 @@ import {
 } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
 import { Subscription, of, BehaviorSubject, combineLatest } from 'rxjs';
-import { debounceTime, take, pairwise, skipWhile } from 'rxjs/operators';
-import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { debounceTime, take, pairwise, skipWhile, first } from 'rxjs/operators';
+import { MatDialog, MatDialogConfig, MatDialogRef } from '@angular/material/dialog';
 import MapBrowserEvent from 'ol/MapBrowserEvent';
 import * as olProj from 'ol/proj';
 import olFeature from 'ol/Feature';
@@ -63,7 +63,8 @@ import {
   WfsWorkspace,
   FeatureWorkspace,
   generateIdFromSourceOptions,
-  computeOlFeaturesExtent
+  computeOlFeaturesExtent,
+  addStopToStore
 } from '@igo2/geo';
 
 import {
@@ -72,7 +73,8 @@ import {
   SearchState,
   QueryState,
   ContextState,
-  WorkspaceState
+  WorkspaceState,
+  DirectionState
 } from '@igo2/integration';
 
 import {
@@ -84,7 +86,7 @@ import {
   mapSlideX,
   mapSlideY
 } from './portal.animation';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 
 import { WelcomeWindowComponent } from './welcome-window/welcome-window.component';
 import { WelcomeWindowService } from './welcome-window/welcome-window.service';
@@ -131,6 +133,7 @@ export class PortalComponent implements OnInit, OnDestroy {
     this.storageService.get('workspaceMaximize') as boolean
   );
 
+  public matDialogRef$ = new BehaviorSubject<MatDialogRef<any>>(undefined);
   public searchBarTerm = '';
   public onSettingsChange$ = new BehaviorSubject<boolean>(undefined);
   public termDefinedInUrl = false;
@@ -307,13 +310,16 @@ export class PortalComponent implements OnInit, OnDestroy {
     private welcomeWindowService: WelcomeWindowService,
     public dialogWindow: MatDialog,
     private queryService: QueryService,
-    private storageService: StorageService
+    private storageService: StorageService,
+    private directionState: DirectionState
   ) {
     this.hasExpansionPanel = this.configService.getConfig('hasExpansionPanel');
     this.hasGeolocateButton =
-    this.configService.getConfig('hasGeolocateButton') === undefined ? true : this.configService.getConfig('hasGeolocateButton') ;
+      this.configService.getConfig('hasGeolocateButton') === undefined ? true : this.configService.getConfig('hasGeolocateButton');
     this.showRotationButtonIfNoRotation =
-    this.configService.getConfig('showRotationButtonIfNoRotation') === undefined ? false : this.configService.getConfig('showRotationButtonIfNoRotation') ;
+      this.configService.getConfig('showRotationButtonIfNoRotation') === undefined ?
+        false :
+        this.configService.getConfig('showRotationButtonIfNoRotation');
     this.forceCoordsNA = this.configService.getConfig('app.forceCoordsNA');
     this.hasFeatureEmphasisOnSelection = this.configService.getConfig(
       'hasFeatureEmphasisOnSelection'
@@ -452,11 +458,16 @@ export class PortalComponent implements OnInit, OnDestroy {
       });
   }
 
+  setToastPanelHtmlDisplay(value) {
+    this.toastPanelHtmlDisplay = value;
+    this.computeToastPanelOffsetX();
+  }
+
   computeToastPanelOffsetX() {
     if (this.isMobile() || !this.isLandscape()) {
-      this.toastPanelOffsetX$.next(undefined);
+      Promise.resolve().then(() => this.toastPanelOffsetX$.next(undefined));
     } else {
-      this.toastPanelOffsetX$.next(this.getToastPanelExtent());
+      Promise.resolve().then(() => this.toastPanelOffsetX$.next(this.getToastPanelExtent()));
     }
   }
 
@@ -596,9 +607,9 @@ export class PortalComponent implements OnInit, OnDestroy {
   }
 
   onSearchTermChange(term?: string) {
-    if (this.routeParams?.search &&  term !== this.routeParams.search) {
+    if (this.routeParams?.search && term !== this.routeParams.search) {
       this.searchState.deactivateCustomFilterTermStrategy();
-  }
+    }
 
     this.searchState.setSearchTerm(term);
     const termWithoutHashtag = term.replace(/(#[^\s]*)/g, '').trim();
@@ -655,6 +666,7 @@ export class PortalComponent implements OnInit, OnDestroy {
 
   private toggleSidenav() {
     this.sidenavOpened ? this.closeSidenav() : this.openSidenav();
+    this.computeToastPanelOffsetX();
   }
 
   public toolChanged(tool: Tool) {
@@ -918,7 +930,9 @@ export class PortalComponent implements OnInit, OnDestroy {
     }
   }
   getControlsOffsetY() {
-    return this.expansionPanelExpanded ? this.workspaceMaximize$.value ? 'firstRowFromBottom-expanded-maximized' : 'firstRowFromBottom-expanded' : 'firstRowFromBottom';
+    return this.expansionPanelExpanded ?
+      this.workspaceMaximize$.value ? 'firstRowFromBottom-expanded-maximized' : 'firstRowFromBottom-expanded' :
+      'firstRowFromBottom';
   }
 
   getBaselayersSwitcherStatus() {
@@ -929,23 +943,23 @@ export class PortalComponent implements OnInit, OnDestroy {
         if (this.expansionPanelExpanded === false) {
           if (this.queryState.store.entities$.value.length === 0) {
             status = 'secondRowFromBottom';
-           } else {
-            status =  'thirdRowFromBottom';
-           }
+          } else {
+            status = 'thirdRowFromBottom';
+          }
         } else {
           if (this.queryState.store.entities$.value.length === 0) {
             status = 'firstRowFromBottom-expanded';
-           } else {
-            status =  'secondRowFromBottom-expanded';
-           }
+          } else {
+            status = 'secondRowFromBottom-expanded';
+          }
         }
 
       } else {
         if (this.queryState.store.entities$.value.length === 0) {
-          status =  'firstRowFromBottom';
-         } else {
-          status =  'secondRowFromBottom';
-         }
+          status = 'firstRowFromBottom';
+        } else {
+          status = 'secondRowFromBottom';
+        }
       }
     } else {
       if (this.workspaceState.workspaceEnabled$.value) {
@@ -971,7 +985,20 @@ export class PortalComponent implements OnInit, OnDestroy {
       this.readToolParams();
       this.readSearchParams();
       this.readFocusFirst();
+      this.computeZoomToExtent();
     });
+  }
+
+  private computeZoomToExtent() {
+    if (this.routeParams['zoomExtent']) {
+      const extentParams = this.routeParams['zoomExtent'].split(',');
+      const olExtent = olProj.transformExtent(
+        extentParams,
+        'EPSG:4326',
+        this.map.projection
+      );
+      this.map.viewController.zoomToExtent(olExtent as [number, number, number, number]);
+    }
   }
 
   private computeFocusFirst() {
@@ -1011,24 +1038,24 @@ export class PortalComponent implements OnInit, OnDestroy {
       }
       if (this.routeParams['search'] && !this.routeParams['zoom'] && this.routeParams['sf'] !== '1') {
         const entities$$ = this.searchStore.stateView.all$()
-        .pipe(
-          skipWhile((entities) => entities.length === 0),
-          debounceTime(500),
-          take(1)
-        )
-        .subscribe((entities) => {
-          entities$$.unsubscribe();
-          const searchResultsOlFeatures = entities
-            .filter(e => e.entity.meta.dataType === FEATURE)
-            .map((entity: EntityRecord<SearchResult>) =>
-              new olFormatGeoJSON().readFeature(entity.entity.data, {
-                dataProjection: entity.entity.data.projection,
-                featureProjection: this.map.projection
-              })
-            );
+          .pipe(
+            skipWhile((entities) => entities.length === 0),
+            debounceTime(500),
+            take(1)
+          )
+          .subscribe((entities) => {
+            entities$$.unsubscribe();
+            const searchResultsOlFeatures = entities
+              .filter(e => e.entity.meta.dataType === FEATURE)
+              .map((entity: EntityRecord<SearchResult>) =>
+                new olFormatGeoJSON().readFeature(entity.entity.data, {
+                  dataProjection: entity.entity.data.projection,
+                  featureProjection: this.map.projection
+                })
+              );
             const totalExtent = computeOlFeaturesExtent(this.map, searchResultsOlFeatures);
             this.map.viewController.zoomToExtent(totalExtent);
-        });
+          });
       }
       this.searchBarTerm = this.routeParams['search'];
     }
@@ -1039,13 +1066,68 @@ export class PortalComponent implements OnInit, OnDestroy {
 
   private readToolParams() {
     if (this.routeParams['tool']) {
-      this.toolbox.activateTool(this.routeParams['tool']);
+      this.matDialogRef$.pipe(
+        skipWhile(r => r !== undefined),
+        first()
+      ).subscribe(matDialogOpened => {
+        if (!matDialogOpened) {
+          this.toolbox.activateTool(this.routeParams['tool']);
+        }
+      });
     }
 
     if (this.routeParams['sidenav'] === '1') {
       setTimeout(() => {
         this.openSidenav();
       }, 250);
+    }
+
+    if (this.routeParams['routing']) {
+      let routingCoordLoaded = false;
+      const stopCoords = this.routeParams['routing'].split(';');
+      const routingOptions = this.routeParams['routingOptions'];
+      let resultSelection: number;
+      if (routingOptions) {
+        resultSelection = parseInt(routingOptions.split('result:')[1], 10);
+      }
+      this.directionState.stopsStore.storeInitialized$
+        .pipe(skipWhile(init => !init), first())
+        .subscribe((init: boolean) => {
+          if (init && !routingCoordLoaded) {
+            routingCoordLoaded = true;
+            stopCoords.map((coord, i) => {
+              if (i > 1) {
+                addStopToStore(this.directionState.stopsStore);
+              }
+            });
+            setTimeout(() => {
+              stopCoords.map((coord, i) => {
+                const stop = this.directionState.stopsStore.all().find(e => e.position === i);
+                stop.text = coord;
+                stop.coordinates = coord.split(',');
+                this.directionState.stopsStore.update(stop);
+              });
+            }, this.directionState.debounceTime * 1.25); // this delay is due to the default component debounce time
+          }
+        });
+      // zoom to active route
+      this.directionState.routesFeatureStore.count$
+        .pipe(skipWhile(c => c < 1), first())
+        .subscribe(c => {
+          if (c >= 1) {
+            this.directionState.zoomToActiveRoute$.next();
+          }
+        });
+      // select the active route by url controls
+      this.directionState.routesFeatureStore.count$
+        .pipe(skipWhile(c => c < 2), first())
+        .subscribe(() => {
+          if (resultSelection) {
+            this.directionState.routesFeatureStore.entities$.value.map(d => d.properties.active = false);
+            this.directionState.routesFeatureStore.entities$.value[resultSelection].properties.active = true;
+            this.directionState.zoomToActiveRoute$.next();
+          }
+        });
     }
   }
 
@@ -1056,6 +1138,15 @@ export class PortalComponent implements OnInit, OnDestroy {
     this.readLayersQueryParamsByType(params, 'imagearcgisrest');
     this.readLayersQueryParamsByType(params, 'tilearcgisrest');
     this.readVectorQueryParams(params);
+  }
+
+  getQueryParam(name, url) {
+    let paramValue;
+    if (url.includes('?')) {
+      const httpParams = new HttpParams({ fromString: url.split('?')[1] });
+      paramValue = httpParams.get(name);
+    }
+    return paramValue;
   }
 
   private readLayersQueryParamsByType(params: Params, type) {
@@ -1100,7 +1191,16 @@ export class PortalComponent implements OnInit, OnDestroy {
     const urls = params[urlsKey].split(',');
 
     let cnt = 0;
-    urls.forEach((url) => {
+    urls.forEach((urlSrc) => {
+      let url = urlSrc;
+      const version =
+        this.getQueryParam('VERSION', url) ||
+        this.getQueryParam('version', url) ||
+        undefined;
+      if (version) {
+        url = url.replace('VERSION=' + version, '').replace('version=' + version, '');
+      }
+
       const currentLayersByService = this.extractLayersByService(
         layersByService[cnt]
       );
@@ -1118,6 +1218,7 @@ export class PortalComponent implements OnInit, OnDestroy {
           url,
           layerFromUrl[0],
           type,
+          version,
           visibility,
           layerFromUrl[1] ? parseInt(layerFromUrl[1], 10) : undefined
         );
@@ -1177,7 +1278,8 @@ export class PortalComponent implements OnInit, OnDestroy {
   private addLayerFromURL(
     url: string,
     name: string,
-    type: 'wms' | 'wmts' | 'arcgisrest'| 'imagearcgisrest' | 'tilearcgisrest',
+    type: 'wms' | 'wmts' | 'arcgisrest' | 'imagearcgisrest' | 'tilearcgisrest',
+    version: string,
     visibility: boolean = true,
     zIndex: number
   ) {
@@ -1194,12 +1296,12 @@ export class PortalComponent implements OnInit, OnDestroy {
     const arcgisClause = (type === 'arcgisrest' || type === 'imagearcgisrest' || type === 'tilearcgisrest');
     let sourceOptions = {
       version: type === 'wmts' ? '1.0.0' : undefined,
-      queryable: arcgisClause  ? true : false,
-      queryFormat: arcgisClause  ? 'esrijson' : undefined,
+      queryable: arcgisClause ? true : false,
+      queryFormat: arcgisClause ? 'esrijson' : undefined,
       layer: name
     };
     if (type === 'wms') {
-      sourceOptions =  { params: {LAYERS: name}} as any;
+      sourceOptions = { params: { LAYERS: name, VERSION: version } } as any;
     }
 
     sourceOptions = ObjectUtils.removeUndefined(Object.assign({}, sourceOptions, commonSourceOptions));
@@ -1249,7 +1351,7 @@ export class PortalComponent implements OnInit, OnDestroy {
     // After, managing named layer by id (context.json OR id from datasource)
     visiblelayers = visibleOnLayersParams.split(',');
     invisiblelayers = visibleOffLayersParams.split(',');
-    if (visiblelayers.indexOf(currentLayerid) > -1  || visiblelayers.indexOf(currentLayerid.toString()) > -1) {
+    if (visiblelayers.indexOf(currentLayerid) > -1 || visiblelayers.indexOf(currentLayerid.toString()) > -1) {
       visible = true;
     }
     if (invisiblelayers.indexOf(currentLayerid) > -1 || invisiblelayers.indexOf(currentLayerid.toString()) > -1) {
@@ -1274,13 +1376,15 @@ export class PortalComponent implements OnInit, OnDestroy {
     if (this.welcomeWindowService.hasWelcomeWindow()) {
       const welcomWindowConfig: MatDialogConfig = this.welcomeWindowService.getConfig();
 
-      const dialogRef = this.dialogWindow.open(
-        WelcomeWindowComponent,
-        welcomWindowConfig
-      );
+      this.matDialogRef$.next(
+          this.dialogWindow.open(
+          WelcomeWindowComponent,
+          welcomWindowConfig
+        ));
 
-      dialogRef.afterClosed().subscribe((result) => {
+        this.matDialogRef$.value.afterClosed().subscribe((result) => {
         this.welcomeWindowService.afterClosedWelcomeWindow();
+        this.matDialogRef$.next(undefined);
       });
     }
   }
